Optimize table performance
