+++ bmh_lims/frontend/src/components/index.js
@@ -8,3 +8,4 @@ export { Notice } from './Modal/'
 export { MultilineText } from './MultilineText/'
 export { Checkbox } from './Checkbox/'
 export { default as CircularButtonBar } from './CircularButtonBar/'
+export { default as DropdownMenu } from './DropdownMenu/'
